Add Probes to driver container daemonsets
This would ensure proper operation of the operator
in regards to dependencies between components as
the network-operator waits for one stage group to be done
before moving the the other.

In addition, liveliness probes would trigger a restart
of the container in case the driver was suddenly unloaded
for some reason.