Remove kernel version dependency from stage ofed driver
This commit removes the dependency of kernel version from
MOFED deployment state.

It is now expected that MOFED driver container to handle
running on a different kernel version than they were compiled
on.

Note:
Doing so will increase the time until a driver container loads
on system startup in case it does not have precompiled modules
for the running kernel, however it will greatly reduce the amount
of containers needed to be built for a certain driver version

Signed-off-by: Adrian Chiris <adrianc@nvidia.com>