Change default Multus CNI version
Due to Issue: https://github.com/intel/multus-cni/issues/592
When multus is removed from the cluster it is left in an inoperable
as CNI calls fail due to removed RBAC.

While it makes sense to occur for secondary network when specified
in a pod spec as net-attach-def name, it should not fail on primary
network.

Downgrade multus CNI to v3.4.1 where such an issue does not occur.

Signed-off-by: Adrian Chiris <adrianc@nvidia.com>