Stop using SelfLink
Objects SelfLink was deprecated in Kubernetes v1.16 and will be
removed in v.1.21. SelfLink is disabled by default in v.1.20.

This implementation is a simplified version of the original one [1]
and uses public API which isn't supposed to be changed soon.

This patch also moves MacvlanNetwork status update to controller.

[1] https://github.com/kubernetes/kubernetes/blob/54449be03127f883e52df48b95adef36925ae2cf/staging/src/k8s.io/apiserver/pkg/endpoints/handlers/namer.go#L105

Closes: #133
Signed-off-by: Ivan Kolodiazhnyi <ikolodiazhny@nvidia.com>