Skip network-operator pod when draining nodes during OFED upgrade
We want to skip network-operator itself during the drain because
the upgrade process might hang if the operator is evicted and
can't be rescheduled to any other node, e.g. in a single-node cluster.
It's safe to do because the goal of the node draining during
the upgrade is to evict pods that might use MOFED
and network-operator doesn't use in its own pod.

Signed-off-by: amaslennikov <amaslennikov@nvidia.com>