Requeue if pod not available in AutoUpgrade
In the auto upgrade flow, the upgrade state of the nodes is built
based on the Ofed pods.

However, there can be a stage where pods are not available as
the Upgrade flow is deleting them and new ones were not created
yet.

In that case, the corresponding node will not be considered in the
upgrade logic at this specific time when the pod is not re-created.

This can affect the "MaxParallelUpgrades" feature.

With this commit, in case a Ofed DaemonSet has one or more unavailable
pods, the upgrade logic will not run and the notification will be requeue
to be handle later.

Signed-off-by: Fred Rolland <frolland@nvidia.com>