Set a default timeout for node drain in OFED auto-upgrade config
When attempting to drain a node, a non-fatal error might occur and
the drain library will retry the action which caused this error
in a few seconds (for example, evicting the pod violates
the deployment's disruption budget).
If the error's cause is not fixed, there will be infinite retries
unless the drain is terminated by timeout.
Since drain operation is disruptive and makes the node unschedulable,
introducing the default timeout to avoid taking the node
out of action indefinitely.

Signed-off-by: amaslennikov <amaslennikov@nvidia.com>