Align the skip-drain label for the operator with the common upgrade lib
When the upgrade functionality was extracted into a separate upgrade
library, the pod label for skipping drain was changed. The operator's
spec was not aligned to use this new label.

This label is required for the operator, when deployed
on a single-node cluster, where operator runs on the same node as
the ofed driver container. Without this label, the operator pod would
drain itself, thus breaking the complete network-operator deployment.

Signed-off-by: amaslennikov <amaslennikov@nvidia.com>