Add default env vars for MOFED container
In case of older distros (kernels),
we need to explictly specify CREATE_IFNAMES_UDEV
so that netdev names are preserved.

when MOFED container loads the driver, it reports
additional information to kernel about Nvidia network
devices that affect the way the OS names netdevs.

On newer distros (kernels) this is not needed as
inbox driver already reports this information.

Instruct MOFED container to install the
"name preserving udev" on the following OSes:

rhel < 9.0
ocp < 4.13
ubuntu < 22.04

Signed-off-by: adrianc <adrianc@nvidia.com>