bug: delete NCP on Helm uninstall
In PR #593, the creation of the NicClusterPolicy CR via Helm
has been marked to be done with Helm "post-install,post-upgrade".

As a result, when Helm was uninstalled, the CR was not deleted.

In order to keep previous behavior, a new post-delete hook is
introduced to delete the CR if it was created by Helm.

Signed-off-by: Fred Rolland <frolland@nvidia.com>