fix: Populate coverage in all packages when generating test coverage
This change ensures that all the packages are taken into account when
generating test coverage for a particular package e.g. integration
tests on controllers end up marking the `Sync()` methods of pkg/state
as covered which is not the case without this change.

The drawback of this change is that unit-test coverage is not clearly
visible as it used to be.

Signed-off-by: Vasilis Remmas <vremmas@nvidia.com>