feat: Allow to configure persistent driver inventory for the OFED driver to avoid recompilation between runs
Introduce DriverInventoryPath to the OFED spec.

It represents a path to a persistent writable folder where OFED container can store its build artifacts. If the artifacts for the correct version and arch exist in this folder when the OFED container is started, they will be reused without driver being recompiled.

Signed-off-by: amaslennikov <amaslennikov@nvidia.com>